// src/modules/work-item-comments/persistence/schemas/work-item-comment.schema.ts
export interface WorkItemCommentDocProps {
  commentPublicId: string
  workspacePublicId: string
  projectPublicId: string
  backlogItemPublicId: string
  body: string
  createdByUserPublicId: string
  createdAt: Date
  updatedAt: Date
  deletedAt: Date | null
  deletedByUserPublicId: string | null
}